remote_server/requirements.txt
# VisionTracker Remote Server — requirements
# Install with: pip install -r requirements.txt

# --- Web Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# --- LLM Inference (choose one or both) ---
# Option 1: llama-cpp-python for GGUF models (recommended for Colab CPU/T4)
llama-cpp-python>=0.2.20

# Option 2: transformers + accelerate for HuggingFace models
# Supports models like microsoft/Phi-4-multimodal, meta-llama/Llama-3.2-Vision, etc.
transformers>=4.36.0
accelerate>=0.25.0
torch>=2.1.0

# --- Tunnel ---
pyngrok>=7.0.0

# --- Image Processing ---
pillow>=10.0.0

# --- Utilities ---
numpy>=1.24.0
pydantic>=2.5.0
